+++ e2e_tests/conftest.py
@@ -46,7 +46,7 @@ def environment():
     user = raw_env.user_username
     service_account_name = raw_env.service_account_name
 
-    project_name, _ = a_project_name(project_slug=fake.slug())
+    project_name = a_project_name(project_slug=fake.slug())
     project_identifier = normalize_project_name(name=project_name, workspace=workspace)
     created_project_identifier = create_project(
         name=project_name,